requirements.txt
# Core Django
Django>=4.1,<5.0
asgiref>=3.7.2
sqlparse>=0.4.4

# Environment variables
django-environ>=0.9.0
dj-database-url>=2.1.0

# Database
psycopg2-binary>=2.9.9

# Server
gunicorn>=21.2.0
whitenoise>=6.2.0

# Django packages
django-widget-tweaks>=1.4.12
django-debug-toolbar>=3.6.0
django-allauth>=0.62.1
djangorestframework>=3.13.1
djangorestframework-simplejwt>=5.3.1

# Auth / Social
social-auth-app-django>=5.4.1
social-auth-core>=4.5.4
PyJWT>=2.8.0

# PDF generation
xhtml2pdf>=0.2.8
reportlab>=3.6.13
Pillow>=10.0.1
pyHanko>=0.21.0
PyPDF3>=1.0.6
lxml>=5.1.0
arabic-reshaper>=3.0.0
python-bidi>=0.4.2
svglib>=1.5.1

# Payment
sslcommerz-lib>=1.0

# Utilities
requests>=2.31.0
python-dateutil>=2.8.2
pytz>=2021.3
qrcode>=7.4.2
